refactor: simplify demo layout rendering
