+++ src/api-viewer-base.ts
@@ -82,7 +82,7 @@ async function renderDocs(
           `
         : html`
             <api-viewer-demo
-              .name=${data.name}
+              .tag=${data.name}
               .props=${data.properties}
               .slots=${data.slots}
               .events=${data.events}
